refactor: move test cases into their own directory
All test files expose a single `all_tests` function.
