+++ src/testcase/bfi.ptx
