+++ src/testcase/cos.ptx
