+++ src/testcase/lg2.ptx
