+++ src/testcase/lg2.rs
@@ -1,32 +1,36 @@
+use rug::Float;
+
 use crate::common::{self, flush_to_zero_f32};
 use crate::cuda::Cuda;
 use crate::test::{self, RangeTest, TestCase, TestCommon};
 use core::f32;
 use std::mem;
 
-pub static PTX: &str = include_str!("sin.ptx");
+pub static PTX: &str = include_str!("lg2.ptx");
 
-pub(crate) fn all_tests() -> Vec<TestCase> {
+const PRECISION: u32 = 64;
+
+pub fn all_tests() -> Vec<TestCase> {
     let mut tests = vec![];
-    for ftz in [false, true] {
-        tests.push(sin(ftz));
-    }
+    // we don't test any input or output that could be flushed
+    tests.push(lg2(false));
     tests
 }
 
-fn sin(ftz: bool) -> TestCase {
-    let test = Box::new(move |cuda: &Cuda| test::run_range::<Sin>(cuda, Sin { ftz }));
+fn lg2(ftz: bool) -> TestCase {
+    let mut tolerance = Float::with_val(PRECISION, -22.6f64);
+    tolerance.exp2_mut();
+    let test = Box::new(move |cuda: &Cuda| test::run_range::<Lg2>(cuda, Lg2 { ftz, tolerance }));
     let ftz = if ftz { "_ftz" } else { "" };
-    TestCase::new(format!("sin_approx{}", ftz), test)
+    TestCase::new(format!("lg2_approx{}", ftz), test)
 }
 
-pub struct Sin {
+pub struct Lg2 {
     ftz: bool,
+    tolerance: Float,
 }
 
-const APPROX_TOLERANCE: f64 = 0.00000051106141211332948885584179164092160363501768f64; // 2^-20.9
-
-impl TestCommon for Sin {
+impl TestCommon for Lg2 {
     type Input = f32;
 
     type Output = f32;
@@ -43,20 +47,20 @@ impl TestCommon for Sin {
         mut input: Self::Input,
         output: Self::Output,
     ) -> Result<(), Self::Output> {
-        fn sin_approx_special(input: f32) -> Option<f32> {
+        fn lg2_approx_special(input: f32) -> Option<f32> {
             Some(match input {
                 f32::NEG_INFINITY => f32::NAN,
-                f if f.is_subnormal() && f.is_sign_negative() => -0.0,
-                f if f.to_ne_bytes() == (-0.0f32).to_ne_bytes() => -0.0,
-                0.0 => 0.0,
-                f if f.is_subnormal() && f.is_sign_positive() => 0.0,
-                f32::INFINITY => f32::NAN,
+                f if f.is_subnormal() && f.is_sign_negative() => f32::NEG_INFINITY,
+                f if f.to_ne_bytes() == (-0.0f32).to_ne_bytes() => f32::NEG_INFINITY,
+                0.0 => f32::NEG_INFINITY,
+                f if f.is_subnormal() && f.is_sign_positive() => f32::NEG_INFINITY,
+                f32::INFINITY => f32::INFINITY,
                 f if f.is_nan() => f32::NAN,
                 _ => return None,
             })
         }
         flush_to_zero_f32(&mut input, self.ftz);
-        if let Some(mut expected) = sin_approx_special(input) {
+        if let Some(mut expected) = lg2_approx_special(input) {
             flush_to_zero_f32(&mut expected, self.ftz);
             if (expected.is_nan() && output.is_nan())
                 || (expected.to_ne_bytes() == output.to_ne_bytes())
@@ -66,26 +70,26 @@ impl TestCommon for Sin {
                 Err(expected)
             }
         } else {
-            let precise_result = sin_host(input);
-            let mut result_f32 = precise_result as f32;
-            flush_to_zero_f32(&mut result_f32, self.ftz);
-            let precise_output = output as f64;
-            let diff = (precise_output - result_f32 as f64).abs();
-            if diff <= APPROX_TOLERANCE {
+            let precise_result = lg2_host(input);
+            let actual_result = Float::with_val(PRECISION, output);
+            let mut diff = precise_result.clone() - actual_result.clone();
+            diff = diff.abs();
+            if diff <= self.tolerance {
                 Ok(())
             } else {
-                Err(precise_result as f32)
+                Err(precise_result.to_f32())
             }
         }
     }
 }
 
-const RANGE_MIN: f32 = 0f32;
-const RANGE_MAX: f32 = f32::consts::FRAC_PI_2;
+const RANGE_MIN: f32 = 1f32;
+const RANGE_MAX: f32 = 4f32;
 
-impl RangeTest for Sin {
+impl RangeTest for Lg2 {
     const MAX_VALUE: u32 =
-        (unsafe { mem::transmute::<_, u32>(RANGE_MAX) - mem::transmute::<_, u32>(RANGE_MIN) }) + 36;
+        (unsafe { mem::transmute::<_, u32>(RANGE_MAX) - mem::transmute::<_, u32>(RANGE_MIN) })
+            + 127;
 
     fn generate(&self, input: u32) -> Self::Input {
         let max_number = unsafe { mem::transmute::<_, u32>(RANGE_MAX) };
@@ -106,7 +110,7 @@ impl RangeTest for Sin {
     }
 }
 
-fn sin_host(input: f32) -> f64 {
-    let input = input as f64;
-    input.sin()
+fn lg2_host(input: f32) -> rug::Float {
+    let input = rug::Float::with_val(PRECISION, input);
+    input.log2()
 }
