+++ src/testcase/minmax.ptx
