+++ src/testcase/shift.ptx
