+++ src/testcase/sin.ptx
